src/app/requirements.txt
flask==3.0.0
databricks-sdk==0.70.0
python-dotenv==1.0.0
